Split the common test classes to outside the messenger test and refactored that to specify queues
