+++ src/Tests/Eshopworld.Messaging.Tests/MessengerQueueTest.cs
@@ -15,11 +15,11 @@ using Xunit;
 // ReSharper disable once CheckNamespace
 // ReSharper disable AccessToDisposedClosure
 [Collection(nameof(AzureServiceBusCollection))]
-public class MessengerTest
+public class MessengerQueueTest
 {
     internal readonly AzureServiceBusFixture ServiceBusFixture;
 
-    public MessengerTest(AzureServiceBusFixture serviceBusFixture)
+    public MessengerQueueTest(AzureServiceBusFixture serviceBusFixture)
     {
         ServiceBusFixture = serviceBusFixture;
     }
@@ -233,42 +233,5 @@ public class MessengerTest
             ((QueueAdapter<TestMessage>)msn.ServiceBusAdapters[typeof(TestMessage)]).Messages.Should().BeEmpty();
             ((QueueAdapter<TestMessage>)msn.ServiceBusAdapters[typeof(TestMessage)]).LockTimers.Should().BeEmpty();
         }
-    }
-}
-
-/// <summary>
-/// A convenient way to generate random Lorem text around a test message implementation.
-/// </summary>
-public class TestMessage : IEquatable<TestMessage>
-{
-    private static readonly Random Rng = new Random();
-
-    public string Name { get; set; }
-
-    public string Stuff { get; set; }
-
-    public float Price { get; set; }
-
-    public TestMessage()
-    {
-        Name = Lorem.GetSentence();
-        Stuff = Lorem.GetParagraph();
-        Price = Rng.Next(100);
-    }
-
-    public TestMessage(string name, string stuff, float price)
-    {
-        Name = name;
-        Stuff = stuff;
-        Price = price;
-    }
-
-    public bool Equals(TestMessage other)
-    {
-        if (other is null) return false;
-        if (ReferenceEquals(this, other)) return true;
-        return string.Equals(Name, other.Name) &&
-               string.Equals(Stuff, other.Stuff) &&
-               Price.Equals(other.Price);
     }
 }
